memperbaiki route, bug rating,tampilan halaman bisnis dan profil bisnis, mengubah struktur database,  js rating, dan 1000 bug lainnya

+++ database/migrations/2021_05_21_134956_create_ratings_table.php
@@ -17,7 +17,7 @@ class CreateRatingsTable extends Migration
             $table->id();
             $table->integer('id_user');
             $table->integer('id_restaurant');
-            $table->integer('rating');
+            $table->float('rating',6,1);
             $table->text('pesan');
             $table->timestamps();
         });
